dpkg-bazel: Add caching of bazel fetch results
Hash the variables that influence the content of orig-tarball and also
add a hash of the debian/rules file to this. Store this hash and the
tarball in DL_DIR and check on the next call of do_fetch_bazel_deps if
the cached archive could be used. This allows true offline build and
reduced unneeded downloads.

Closes #2
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>