recipes-devtools/bazel-toolchains: use do_prepare_build
Instead of adding a custom task to create local_toolchains(), use
the existing Isar do_prepare_build task since that's the official
way of making changes to the source tree after do_patch but before
the build (including installation of build dependencies) is scheduled
to happen. This is needed to prepare for sbuild as do_install_buildeps
is removed.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>